agent: simplify task manager
Previously, the task manager handled removal. With the addition of the
worker, we are going to a secondary reaper that will take care of
container cleanup. This has allowed use to revise the task manager
response loop to simply handled outstanding operations and updates,
making the behavior much less reliable.

We've also greatly increased the unit test coverage, albeit the tests
are minimal and non-exhaustive.

Signed-off-by: Stephen J Day <stephen.day@docker.com>