agent: initial sketch of worker
We've now made enough changes to define the Worker interface. The Worker
accepts repeated calls to maintain task assignments to a node. Tasks
inside the set are run and reported. Those outside the set are removed
and evicted from the node's storage.

This initial sketch attempts at on disk storage but still requires
testing and verification.

Changes were made to the status reporter to support this addition.

Signed-off-by: Stephen J Day <stephen.day@docker.com>